fiexed StakeExpenditureDialog
