+++ src/modules/dashboard/components/Dialogs/StakeExpenditureDialog/index.ts
@@ -1,2 +1,1 @@
 export { default } from './StakeExpenditureDialog';
-export { default as StakeExpenditureDialogForm } from './StakeExpenditureDialogForm';
